pyproject.toml
[tool.poetry]
name = "my-bot"
version = "0.1.0"
description = ""
authors = ["a.pushkin <a.v.pushkin@gmail.com>"]
readme = "README.md"
packages = [{include = "my_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
pyrogram = {extras = ["fast-mode"], version = "^2.0.106"}
aiohttp = "^3.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
